finance_project/requirements.txt
fastapi
uvicorn
beanie
motor
yfinance
certifi
python-dotenv
